apex: refactor driver initialisation
Drivers used to be initialised by using some linker magic to assemble
and sort an initialisation list.

That approach doesn't scale if you want to be able to reuse drivers
across SOCs/boards as each driver needs to know intimiate details about
the hardware it's running on.

Refactor to have an explicit machine_driver_init hook to initialise
drivers for a particualr piece of hardware.

Long term it would be nice to take something like a device tree and
auto-generate the machine_driver_init function at compile time.