sys/lib/jhash: support unaligned data for jhash & jhash_string
Previously this was a bit broken. It's really not valid to cast a void*
to another pointer if you have no idea about alignment. And that's
before you consider aliasing violations.

Using a memcpy is a good way to load from unaligned. It lets gcc work
out the best way to do it. Alternatively we could rewrite the functions
in terms of char* pointers.