boot: support sparse kernel memory
Previously, the kernel memory reservation was passed to the kernel as a
single bootinfo entry.

Some machines have sparse memory (e.g. tightly coupled memories on ARM
SOCs). To support loading parts of the kernel to different memory
regions we need to split the kernel memory reservations and pass them to
the kernel individually.