apex: properly unwind terminated threads
Previously threads were abruptly terminated without having a chance to
unlock mutexes or release state.

We now send a special signal to a thread to ask it to quit. The thread
is then marked as exiting and will continue to run intil it calls
sch_exit().

While the thread is exiting it cannot take any locks or go to sleep
(exactly the same as receiving a signal).

The kernel will now need an audit to make sure that all paths properly
check the return codes of mutex_lock, sch_sleep, etc.