apex: fix startup file descriptor race condition
Previously exec_into always resumed the main thread it created for the
new task. This was problematic as it was then impossible to correctly
initialise the file descriptors for init.

This change makes exec_into return a pointer to the main thread. Callers
are then expected to call thread_resume on on this pointer when they are
ready.

Note that the interface returns an invalid (but not null) pointer on
error containing the negative error code.