sys/mem/page: properly handle interruptible locks
In APEX all locks & sleeps are interruptible except for the global
scheduler lock.

Handle lock failures due to interruption and return appropriate error
codes from page allocation functions.

Note that page_free is special: if it attempts to lock a region and
fails it falls back to using a the global lock. Therefore, if page_free
is passed valid arguments it is guaranteed to succeed.