mk: define *_rule_exists for all valid targets
Sometimes it is helpful to conditionally disable parts of a makefile if
a particular target has no rule.

For example, for some targets we need to wrap apeximg in some extra
headers in order to make it bootable. However, apeximg is optional.

Unfortunately this approach is error prone as it is dependent on the
order which makefiles are processed, which explains the hackery in
configure. This can definitely be improved.