sys/kern/sch: improve zombie thread reaping
Previously, a thread could be leaked because code after the call to
context_switch() in sch_switch() does not run when restoring the
context of a new thread.

Furthermore, calling thread_free from the depths of the scheduler was a
bit short-sighted as thread_free calls kmem_free which in turn may need
to acquire/release locks causing confusing recursion within the
scheduler.

Fix all of this by adding zombies to a list which is later pruned at an
appropriate point in sch_unlock_slowpath.