apex: further cleanup task exit
We can't release task resources until we know that all threads in the
task have been terminated.

Note that fs_exit is special as it needs to close files as soon as the
last thread in a task exits to unblock waiters in other tasks.

Note also that this will require another round of refactoring as we
currently assume that all processes will be waited on (and therefore
must become zombies). If SIGCHLD is ignored or SA_NOCLDWAIT is set the
task will be leaked.