arch/arm/v7m: fix syscall restart after multiple signals
If multiple signals were delivered immediately one after another the
return address in the exception frame was adjusted multiple times.

This would lead to register corruption on syscall restart or crashes in
userspace.

Fix by adjusting the return address once at the point we know we're
actually returning to userspace.