arch/arm/v7m: fix register corruption after signal delivery from PendSV
If a signal was delivered during return from PendSV exception
sc_sigreturn would previously restore r4-r7 from garbage stack values
and corrupt the exception frame.

r4-r7 are preserved so that non-volatile syscall arguments in r4, r5 and
r7 can be restored for syscall restart.

The original exception frame needs to be preserved in this case as we
are returning from some other exception, not a SVCall.