sys/kern/exec: support PT_GNU_STACK segment
This patch adds support for the PT_GNU_STACK segment header. We now
support:
* Stack attributes (no-execute)
* Stack size for initial thread controlled by segment header

APEX requires programs to specify an initial stack size and does not
fall back to an arbitrary default value. This can be specified by using
the "-z stack-size=xx" argument to ld.

Programs which do not specify this value will fail with ENOEXEC.