sys/mem/page: trim excess from beginning rather than end of allocations
If an allocation is non power-of-two sized we need to decide whether to
trim the excess pages from the beginning or the end of the power-of-two
aligned region.

Previously, we trimmed from the end. This had the nasty side effect of
causing both ends of a stack to end up with poor alignment once a guard
page was setup.

Trimming from the beginning gives much nicer alignment for stacks.