arch/arm/v7m: userspace stack must always be fully mapped
Previously, we were attempting to dynamically map the userspace stack.

This is not possible on ARMv7-M as the core cannot write an exception
frame to unmapped memory. Due to the hard wired exception entry
behaviour certain registers (e.g. LR) are guaranteed to be modified by
the time the exception handler runs. This means that if the exception
frame could not be written the original processor state cannot be
restored.