arch/arm/v7m: fix handling of instruction fault across region boundary
On ARMv7-M an instruction access can cross a region boundary as
instructions must be word aligned but can be word/dword in length.

As the MPU does not provide a fault address on instruction access
violation we infer the fault address from the return address in the
exception frame. However, this is not actually the fault address but
rather the address of the instruction. The fault address can actually
be two bytes further ahead.

To work around this we now pass the maximum access length to mpu_fault
and insert two regions when necessary.

Note that without inspecting the state of the MPU it is impossible to
determine whether we are running off the end of a region or branching to
the middle of a new one. Because of this we will end up with a
duplicated region for the former. Due to the design of the MPU this is
harmless and the resultant code is faster, shorter and much easier to
understand than if we tried to avoid the duplicate.