libc++: fix alignment of std::atomic
The C11 atomic functions used as the backend of std::atomic require
their data to be naturally aligned.

There are discussions on this topic upstream, e.g.:
https://bugs.llvm.org/show_bug.cgi?id=26462

Once an upstream solution is available we can use it. For now, let's take
the GCC approach and naturally align power-of-2 sized data up to 16
bytes.