configure: reduce driver initialisation stack use
The lifetime of a compound literal in C is tied to the enclosing block,
so the previous implementation would always use lots of stack.

Theoretically we should be able to encase each driver initialisation
call in its own block and avoid the issue, however gcc versions before
gcc 9 extend the lifetime of a compound literal to the enclosing
function[1].

For now, we can work around the issue by creating an explicit temporary.

[1] https://gcc.gnu.org/gcc-9/porting_to.html