apex: support NXP MIMXRT1060-EVK board
The MIMXRT1060-EVK board is almost the same as the MIMXRT1050-EVKB
board with a few pinout changes, a 1060 CPU and QSPI boot by default.

For now we reuse the majority of the MIMXRT1050-EVKB board support. This
can diverge as necessary in the future.

This patch also adds support for QSPI boot on both boards.