sys/dev/usb: simplify device destruction
Previously tty_destroy would return an error if it was in use. This
caused extra complexity through the usb stack as it meant that once a
function was configured it could not necessarily be destroyed.

Changes in device management & tty mean that we can now destroy a
device as long as we are prepared to sleep waiting for any outstanding
operations to be completed.

This greatly simplifies device destruction in the usb stack and fixes
some race conditions in gadget configuration.