apex: merge devfs & device layer
This change is a small step towards unifying devices & the file system
code, folling the "everything is a file" philosophy.

The reason to take this partial step now is twofold:
1. Previously it was impossible to destroy an in-use device
2. The VFS used to call drivers while holding the vnode lock

In order to fix point 1 we need to also address point 2.

For now we do some hacking to release the vnode lock while calling
driver functions. This will be cleaned up in a future filesystem
rewrite.