apex: replace bootinfo with kernel arguments
Bootinfo was an inflexible fixed structure which did not allow for
machine specific context to be passed from the boot loader to the
kernel.

Replacing bootinfo with simple kernel arguments reduces the coupling
between the boot loader & the kernel.

The kernel now parses it's own ELF headers to determine which memory
regions to reserve.

This will be followed by changes which remove the remnants of struct
bootinfo from the kernel.