sys/kern/sch: remove wake queue
Previously the intention was that waking threads moved to the wake queue
under interrupt, then moved from the wake queue to the run queue at some
point later.

The implementation required an interrupt protected wake queue and
complex locking and flushing logic thoughout the scheduler.

It is much simpler to directly move threads to the run queue.