sys/kern/sch: switch threads under architecture control
This patch is a major rework of the internals of the scheduler to enable
thread switching under architecture control.

When the new arch_schedule() function is called the architecture must
call sch_switch() either immediately or as soon as interrupts are
enabled. This allows an architecture to use an interrupt to switch
threads.