apex: add u_access_suspend/u_access_resume
This is a short term hack to work around problems caused by having a
single lock on the process address space.

For example, if a thread A calls read on a tty and blocks waiting for
input the address space read lock is held. If thread B then tries to
mmap some memory it needs to block until thread A releases the address
space read lock. This is far from an ideal situation.

The u_access_suspend/u_access_resume interface allows the tty driver to
temporarily release the address space lock while waiting for input.

A far better solution to this problem is to selectively lock individual
pages. Unfortunately I don't have time to do that right now.