sys/arch/arm/v7m/mpu: disable region before adjusting
Previously we blindly modified a region base address when faulting in a
new MPU region. This can cause some serious issues as the new base
address may not be valid for the existing region size.

According to the ARMv7-M architecture reference manual:

Software must ensure that the value written to the ADDR field aligns
with the size of the selected region.

By disabling the region before adjusting the base address we can meet
this requirement.