regulator: change supports() interface to return bool
Returning an error code from supports() led to incorrect code in the mmc
framework and the gpio regulator.

The mmc framework assumed positive 'if (supports())' logic. The gpio
regulator returned a boolean result instead of an error code.

Returning bool leads to more natural code and only required the fixed
regulator to be updated.