sch: fix sch_{prepare,continue}_sleep preemption race
9402926eb2a358142893df8eac09ad1eabd1f094 changed to keeping preemption
enabled between sch_prepare_sleep and sch_continue_sleep. This was
incorrect as it allows active_thread to be preempted after it has been
added to the sleep queue which means that it will not run again until a
wakeup.

Disabling preemption guarantees that active_thread will run until
it calls sch_continue_sleep or sch_cancel_sleep.