syslog: fix syslog_format stalling on long message
Previously syslog_format would continuously wait for a bigger buffer
(which never arrived) if a message didn't fit. This effectively stalled
the output.

I guess the idea used to be that a consumer might be able to provide a
bigger buffer, but that's probably fairly difficult in a panic scenario.

Change to truncating output with an ellipsis.