sys/mem/vm: combine contiguous segments
mmap and mprotect modify the address space for a task by either
inserting new segments or splitting/adjusting existing segments.

mmap (via as_insert and then seg_extend) used to keep the address space
defragmented if a new segment followed an existing segment. This
generally meant that sbrk would keep extending a segment until the
allocated pages became discontiguous.

The general solution to this is to walk the segment list and combine
adjacent contiguous segments if their properties are compatible.