sys/dev/fsl/lpuart: reduce RX FIFO watermark value
The watermark value sets how many datawords are
required in the FIFO before an interrupt is generated.
The number of datawords in the FIFO must be greater than
this number.

With the value set to FIFO max - 1, the interrupt is
generated when the FIFO is full. This may lead to
data overruns if the interrupt latency of the system
is greater than 1 character width.

Setting the watermark to a more conservative value
generates the interrupt earlier allowing for more data
to arrive while the interrupt is serviced.