sig: don't force signal handlers to specify a restorer
There's no reason for requiring a signal handler to provide a restorer
other than broken assumptions in the kernel.

Currently thread priority is elevated for signal delivery (for some
historic reason?). The restorer used to reset priority, but doesn't
anymore.

This leads to a bug where once a signal is delivered to a thread it
permanently runs at PRI_SIGNAL. That will be fixed in a subsequent
commit.