sys/fs/vfs: work around calls to fs_exit with preemption disabled
This is a horrible temporary workaround.

Some places in the kernel still use sch_lock() and sch_unlock() for
synchronisation. This caused an assertion in fs_exit as it tries to lock
a mutex. Locking a mutex with preemption disabled is not possible.

Work around this in the same way we're handling fs_exit from interrupt.