libc: update CFLAGS to match musl 1.2.0
Instead of overriding the CFLAGS from the kernel build, set CFLAGS for
musl to something close to the standard musl build.

One departure from the musl build is that we don't specify -nostdinc. At
this time Apex expects to be building against a musl toolchain and
uses the installed C library headers.

Changing this would require a separate bootstrapping step where we build
& install the C library before building the kernel.