configure: statically allocate driver descriptions
Previously driver descriptions were allocated on the stack. This meant
that these descriptions had to be built on or copied to the stack.

This can be slightly more efficient for small descriptions, but quickly
bloats the initialisation code as descriptions grow larger.

The configuration syntax supports function-style initialisation of
drivers which will use stack and register arguments.

Statically allocating descriptions places them in the .rodata section at
compile time.

Note that a driver should not assume that a description will stay valid
after returning from its *_init() function as we will probably unmap
driver descriptions after initialisation at some point in the future.