imxrt/ocotp: replace std::array with builtin array
std::array only guarantees that it can be aggregate initialised, not
that it takes the same amount of space as a builtin array.

While this was not to causing any problems and was guarded by a static
assertion, it is still better practise not to rely on the internal
implementation of std::array as the layout _matters_ in the context of a
register memory map.