fsl/usb2-udc: fix queue reset race in v_stop
Previously we used to clear the run/stop bit in the controller, issue a
controller reset and then reset the queues.

This sequence is problematic as the transition from run to stop and the
controller reset both take time to complete, meaning that the controller
could still be accessing the queues we are about to reset.

I have not found a documented way to determine when the controller has
stopped when running in device mode (the USBSTS.HCH bit is only valid in
host mode).

We can fix this by waiting for the reset to complete before resetting
the queues. Asserting USBCMD.RST also clears UBSCMD.RS, so there is no
need to clear it separately.