fsl/usb2-udc: fix handling of setup request on primed endpoint
The UDC layer was very rarely reporting:

	udc::setup_request: BUG transaction in progress!

This would lock up the stack and thereby prevent further communications
until a bus reset.

This was caused by mishandling of a setup request on a primed endpoint.

The i.MX RT1060 Reference Manual has a few pertinent things to say on
this topic:

1)	After priming the packet, the DCD must verify a new setup packet
	has not been received by reading the USB.ENDPTSETUPSTAT register
	immediately verifying that the prime had completed. A prime will
	complete when the associated bit in the Endpoint Prime
	(USB_nENDPTPRIME) register is zero and the associated bit in the
	Endpoint Status (USB_nENDPTSTAT) register is a one. If a prime
	fails, ie. The Endpoint Prime (USB_nENDPTPRIME) bit goes to zero
	and the Endpoint Status (USB_nENDPTSTAT) bit is not set, then
	the prime has failed. This can only be due to improper setup of
	the dQH, dTD or a setup arriving during the prime operation. If
	a new setup packet is indicated after the ENDPTPRIME bit is
	cleared, then the transfer descriptor can be freed and the DCD
	must reinterpret the setup packet.

2)	Should a setup arrive after the data stage is primed, the device
	controller will automatically clear the prime status (Endpoint
	Status (USB_nENDPTSTAT)) to enforce

3)	After receiving a new setup packet the status and/or handshake
	phases may still be pending from a previous control sequence.
	These should be flushed & deallocated before linking a new
	status and/or handshake dTD for the most recent setup packet.

Statement 1) leaves out the important detail that attempting to prime an
endpoint while the corresponding ENDPTSETUPSTAT flag is set will fail.
The setup request does not need to arrive during the prime operation,
any time before priming will do. Our driver does not need to test
ENDPTSETUPSTAT after priming as this will immediately trigger an
interrupt.

Statements 1) and 2) should mean that 3) is impossible, however in case
I've misinterpreted something the driver will flush the relevant
endpoints when a setup request is received before resetting the
associated queues.

Previously the driver didn't cater for statement 2) and was expecting
the endpoint to remain primed.