usb/gadget/cdc_acm: queue larger transfers
This is a bit hacky but reduces interrupt load for large transfers by
PAGE_SIZE / 2 / max_packet, which is a factor of 4 for PAGE_SIZE 4KiB
and max_packet 512B.

Ideally we would be able to query the transmit buffer size from the tty.
It is currently hard coded as one page and is not adjustable.

Limiting the transfer size to half the transmit buffer size means that
the driver prepares the next transfer while the current transfer is
processing. This keeps the hardware primed and improves the overall
transfer rate.