fs: convert task filesystem lock to rwlock
In general the filesystem state of a task changes much less frequently
than access to file descriptors. This maps well to an rwlock.

Previously a mutex was used here. This led to poor performance under
some workloads as contention on any vnode could cause the task
filesystem state to remain locked for unacceptably long periods.

Using an rwlock here is still working around poor design in the
filesystem. Unfortunately available time does not permit a rewrite of
this area right now.