apex: improve address space locking
Userspace access is now flagged in the thread state. This means that
callers no longer need to track whether access was suspended and can use
a much simpler u_access_continue interface when processing iovs.

This still isn't a particularly elegant solution to this problem, but
it's hopefully a step in the right direction.