imxrt/gdbinit: stop TRACE32 modifying target registers while running
TRACE32 can show live versions of registers while the target is running.
This is problematic as updating the display can cause writes to target
registers.

For example, in order to display the MPU registers the MPU RNR register
needs to be set. If this happens in the middle of a MemManage exception
the MPU state is corrupted.