devfs: fix incorrect use of vnode reference
Previously devfs would call vref on the vnode for a device during
lookup. The VFS provides an inactive hook which is called when a vnode
is no longer referenced. Use this to clear the vnode reference instead.

This means that when a device is closed the associated vnode can now be
released.