vfs: fix struct file reference counting
This patch changes the meaning of the f_count member in struct file.

Previously f_count counted the number of times a file was referenced in
the file array of tasks. Now it also counts the number of times a file
is in use in a file system operation.

This fixes a bug where a file could be closed while in use thereby
freeing the associated struct file and leading to a crash.

This bug was possible as a filesystem (e.g. devfs) is free to release
the vnode lock to support concurrent operations.

Really this is just another hack on the VFS to keep things functional
until this whole thing is eventually rewritten.