fsl/usdhc: fix busy wait race condition
Previously the transfer complete interrupt was enabled at all times.

This led to a race condition where the previously issued command could
complete and release DAT0 (generating an interrupt) and cause a wakeup
on the thread waiting for the current command to complete.

Now we enable the transfer complete interrupt only when it is required:
* Waiting for a data transfer to complete
* Waiting for DAT0 to be released