elf: rework program loading
The previous ELF loading code had a serious bug where file data would be
loaded into the bss of the program if the elf file was not truncated (by
sstrip) to finish precisely at the end of the data segment.

This rewrite makes less assumptions about the layout of the program
segments tries to be more generic when loading them.

On nommu we don't currently unmap holes in the program image. Doing so
would require a way to iterate the program segments which currently
doesn't exist.

We still assume that the data segment is the last PT_LOAD segment in the
program image.