apex: fix global constructor execution
clang (correctly) decided that the two declared arrays cannot overlap.

Turns out that this code was actually quite wrong. __init_array_start is
a pointer to the base address of an array of function pointers in the
program image and must be treated as such.

The best way to fix this is to follow a good example, in this case
libc_start_init from musl.