apex: fix CONFIG_KSTACK_CHECK stack corruption on gcc riscv32
__builtin_frame_address(0) nominally returns the address of the first
word pushed onto the stack, but this is machine & compiler dependent.

Previously we would initialise the stack with 0xaa up to this point,
which would clobber the stack frame on 32-bit RISCV when using gcc
11.2.0.

There doesn't appear to be a way to determine the size of the stack
frame, so, just leave another 128 bytes untouched which will hopefully
avoid any stack corruption.

Change-Id: I1c0fb619ab98c5612f41fe352f3bb5bd85ebc0e3