apex: make bootargs static
Previously bootargs was on the stack. Passing bootargs as a pointer to
stack memory could lead to corruption of bootargs as the the stack
location can be reused.

Ideally we would place bootargs on the stack for the boot thread (like
args for a userspace thread), but the infrastructure to do so is not yet
available.

Change-Id: Idd55f866c78a0d57718647d1478ad2a324e9f0aa